role segregation was added

+++ internal/services/book_service.go
@@ -1,6 +1,7 @@
 package service
 
 import (
+	"errors"
 	"gorm.io/gorm"
 	"rest-project/internal/models"
 	"time"
@@ -39,12 +40,21 @@ func (s *BookService) GetBookByID(id int) (*models.Book, error) {
 	return s.repo.GetById(id)
 }
 
+func (s *BookService) GetUserBooks(userID uint) ([]models.Book, error) {
+	var user models.User
+	if err := s.db.Preload("Books").First(&user, userID).Error; err != nil {
+		return nil, err
+	}
+	return user.Books, nil
+}
+
 func (s *BookService) Create(userID uint, title, author, publishedAt string, pages int) (*models.Book, error) {
 	book := &models.Book{
 		Title:       title,
 		Author:      author,
 		Pages:       pages,
-		PublishedAt: time.Now(), // можно парсить publishedAt, если хочешь
+		PublishedAt: time.Now(),
+		CreatorID:   userID, //
 	}
 
 	err := s.repo.Create(book)
@@ -78,14 +88,28 @@ func (s *BookService) AddBookToUser(userID uint, bookID uint) error {
 	return s.db.Model(&user).Association("Books").Append(&book)
 }
 
-func (s *BookService) Update(id int, bookEdit *models.BookEdit) (*models.Book, error) {
-	err := s.repo.Update(id, bookEdit)
+func (s *BookService) Update(id int, bookEdit *models.BookEdit, userID uint) (*models.Book, error) {
+	book, err := s.repo.GetById(id)
+	if err != nil {
+		return nil, err
+	}
+	if book.CreatorID != userID {
+		return nil, errors.New("you can only update your own books")
+	}
+	err = s.repo.Update(id, bookEdit)
 	if err != nil {
 		return nil, err
 	}
 	return s.GetBookByID(id)
 }
 
-func (s *BookService) DeleteBook(bookID int) error {
+func (s *BookService) DeleteBook(bookID int, userID uint) error {
+	book, err := s.repo.GetById(bookID)
+	if err != nil {
+		return err
+	}
+	if book.CreatorID != userID {
+		return errors.New("you can only delete your own books")
+	}
 	return s.repo.Delete(bookID)
 }
